+++ app/home/historical/page.tsx
@@ -1,6 +1,7 @@
 import { fetchLighthouses, getHistoricalWeather} from "@/app/lib/data";
 import TableRow from "@/app/ui/historical/dailyTable";
-import LineGraph from "@/app/ui/historical/lineChart";
+import LineGraph from "@/app/ui/historical";
+import { LineChart } from "recharts";
 
 export default async function Page() {
   const lighthouses = await fetchLighthouses();
@@ -9,15 +10,15 @@ export default async function Page() {
       <h1>Historical Weather</h1>
       {lighthouses.map(async(lighthouse) => {
         const { daily } = await getHistoricalWeather(lighthouse.latitude, lighthouse.longitude, 'Europe/Dublin' );
-        console.log("Current....",daily);
+        // console.log("Current....",daily);
         // console.log(lighthouse.name);
 
         return (
           <>
             <h1>{ lighthouse.name }</h1>
             <div>  
-            <LineGraph {...daily}/>
-              </div>
+              <LineGraph daily = {daily}/>
+            </div>
             {/* <div className="bg-white-700 mx-auto my-5 w-[40%] h-[40%]">
                     <Map 
                     id={lighthouse.id}