+++ app/home/page.tsx
@@ -1,17 +1,18 @@
 import { fetchLighthouses } from "../lib/data";
-import Map from '@/app/ui/home/map';
+import Map from '@/app/ui/home';
 
 export default async function Page() {
 
     const lighthouses = await fetchLighthouses();
-    console.log(lighthouses);
+    // console.log(lighthouses);
 
     return (
         <>
-        {lighthouses.map((lighthouse) => (
+        {lighthouses.map((lighthouse,index) => (
             <div className="bg-white-700 mx-auto my-5 w-[40%] h-[40%]">
-                <Map 
-                 id={lighthouse.id} 
+                <Map
+                 key={index}
+                 id={lighthouse.id}
                  name={lighthouse.name} 
                  latitude={lighthouse.latitude} 
                  longitude={lighthouse.longitude}