+++ app/ui/home/index.tsx
@@ -3,7 +3,7 @@ import dynamic from "next/dynamic";
 
 
 const Map = dynamic(
-    () => import('@/app/ui/home/map/map'),
+    () => import('@/app/ui/home/map'),
     { 
       loading: () => <p>A map is loading</p>,
       ssr: false