+++ app/ui/home/nav-links.tsx
@@ -13,7 +13,7 @@ import { usePathname } from 'next/navigation';
 const links = [
   { name: 'Home', href: '/home', icon: HomeIcon },
   { name: 'Forecast',href: '/home/forecast',icon: DocumentDuplicateIcon,},
-  { name: 'Marine', href: '/home/forecast/marine', icon: UserGroupIcon },
+  { name: 'Marine', href: '/home/marine', icon: UserGroupIcon },
   { name: 'Historical', href: '/home/historical', icon: UserGroupIcon },
 ];